refactor: moves component and story book file to sub directory
Relates #422
