+++ components/messaging/ConversationList/ConversationsList.tsx
@@ -7,8 +7,8 @@ import {
 } from '@/types/messagingTypes';
 import { useConversationContext } from '@/context/conversationContext';
 import updateConversationReadStatus from '@/supabase/models/messaging/updateConversationReadStatus';
-import ConversationCard from './ConversationCard/ConversationCard';
-import ConversationFilter from './ConversationFilter/ConversationFilter';
+import ConversationCard from '../ConversationCard/ConversationCard';
+import ConversationFilter from '../ConversationFilter/ConversationFilter';
 import conversationWatcher from '@/supabase/channels/conversationListWatcher';
 
 const ConversationsList: React.FC = () => {
